Relation Include: Create own type, pipe and related exceptions
