+++ src/library/models/library.query-parameters.ts
@@ -4,6 +4,7 @@ import type OmitId from "src/utils/omit-id";
 import type OmitSlug from "src/utils/omit-slug";
 import type RequireOnlyOne from "src/utils/require-only-one";
 import { buildStringSearchParameters, SearchStringInput } from "src/utils/search-string-input";
+import type { RelationInclude as BaseRelationInclude } from "src/relation-include/models/relation-include" ;
 
 namespace LibraryQueryParameters {
 
@@ -57,7 +58,7 @@ namespace LibraryQueryParameters {
 	 * The relation field to include in a returned library
 	 */
 	export const AvailableIncludes = ['files'] as const;
-	export type RelationInclude = Record<typeof AvailableIncludes[number], boolean>;
+	export type RelationInclude = BaseRelationInclude<typeof AvailableIncludes>;
 	/**
 	 * Build the query parameters for ORM to include relations
 	 * @returns the ORM-ready query parameters
