+++ src/relation-include/models/relation-include.ts
@@ -1,0 +1,4 @@
+/**
+ * A record to request relations to be included in an object from the database
+ */
+export type RelationInclude<T extends readonly string[]> = Record<T[number], boolean>;
